cpp change

+++ lib/task18/cpp_impl/triangle.cpp
@@ -3,8 +3,8 @@
 using namespace std;
 
 static int solve_triangle(vector<vector<int>> triangle) {
-    for (int row = triangle.size() - 2; row >= 0; row--) {
-        for (int col = 0; col < triangle[row].size(); col++) {
+    for (int row = (int)triangle.size() - 2; row >= 0; row--) {
+        for (int col = 0; col < (int)triangle[row].size(); col++) {
             triangle[row][col] += max(triangle[row + 1][col], triangle[row + 1][col + 1]);
         }
     }
